REFACTOR: Mitigate cyclic dependency between Jsonable classes
