+++ src/PersonObjects/Player/PlayerObject.ts
@@ -22,7 +22,7 @@ import { HashManager } from "../../Hacknet/HashManager";
 import { type MoneySource, MoneySourceTracker } from "../../utils/MoneySourceTracker";
 import { constructorsForReviver, Generic_toJSON, Generic_fromJSON, IReviverValue } from "../../utils/JSONReviver";
 import { JSONMap, JSONSet } from "../../Types/Jsonable";
-import { cyrb53 } from "../../utils/StringHelperFunctions";
+import { cyrb53 } from "../../utils/HashUtils";
 import { getRandomIntInclusive } from "../../utils/helpers/getRandomIntInclusive";
 import { CONSTANTS } from "../../Constants";
 import { Person } from "../Person";
