+++ src/utils/StringHelperFunctions.ts
@@ -89,26 +89,6 @@ export function generateRandomString(n: number): string {
   return str;
 }
 
-/**
- * Hashes the input string. This is a fast hash, so NOT good for cryptography.
- * This has been ripped off here: https://stackoverflow.com/a/52171480
- * @param str The string that is to be hashed
- * @param seed A seed to randomize the result
- * @returns An hexadecimal string representation of the hashed input
- */
-export function cyrb53(str: string, seed = 0): string {
-  let h1 = 0xdeadbeef ^ seed;
-  let h2 = 0x41c6ce57 ^ seed;
-  for (let i = 0, ch; i < str.length; i++) {
-    ch = str.charCodeAt(i);
-    h1 = Math.imul(h1 ^ ch, 2654435761);
-    h2 = Math.imul(h2 ^ ch, 1597334677);
-  }
-  h1 = Math.imul(h1 ^ (h1 >>> 16), 2246822507) ^ Math.imul(h2 ^ (h2 >>> 13), 3266489909);
-  h2 = Math.imul(h2 ^ (h2 >>> 16), 2246822507) ^ Math.imul(h1 ^ (h1 >>> 13), 3266489909);
-  return (4294967296 * (2097151 & h2) + (h1 >>> 0)).toString(16);
-}
-
 export function capitalizeFirstLetter(s: string): string {
   return s.charAt(0).toUpperCase() + s.slice(1);
 }
